requirements.txt
Django==4.2.7
Pillow==10.1.0
celery==5.3.4
redis==5.0.1
requests==2.31.0
python-dotenv==1.0.0
